# swagger/swagger.yaml
swagger: "2.0"
info:
  version: '1.0.0'
  title: Swagger sam-api
...
  contact:
    email: norbert.bondarczuk@wipro.com
host: localhost:8000
basePath: /api
securityDefinitions: {}
schemes:
- http
consumes:
- application/json
produces:
- application/json
paths:
...
  /release/new:
    post:
      description: "Produces a new release of version in state C like Controlled to P like Production allocating a new release id. The impact is on Account and Order entries. Existence of not controled entries with status W like Working causes failure. Other validation rules are applied as well, for example valid date check: it must be rounded down and in the future and no clash with entries in status P like Production found in Account or Order backednd tables. \n\nRequires:\n- Control role."
      summary: ReleaseNew
      tags:
      - release
      operationId: ReleaseNew
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        type: string
        format: uuid
        description: ''      
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/ResultSetOk'
          headers: {}
        401:
          description: Not authenticated
          schema:
            $ref: '#/definitions/ResultSetError'
        500:
          description: Server error
          schema:
            $ref: '#/definitions/ResultSetError'
  /release/{release}:
    post:
      description: "Does a release of version in state C like Controlled to P like Production augmenting target version. The impact is on Account and Order entries. Existence of not controled entries with status W like Working causes failure. Other validation rules are applied as well, for example valid date check: it must be rounded down and in the future and no clash with entries in status P like Production found in Account or Order backednd tables. \n\nRequires:\n- Control role."
      summary: ReleaseAppend
      tags:
      - release
      operationId: ReleaseAppend
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        type: string
        format: uuid
        description: ''
      - name: release
        in: path
        required: true
        enum:
        - 0
        - last
        type: string
        description: release sequential number, 0 for work or control, else production      
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/ResultSetOk'
          headers: {}
        401:
          description: Not authenticated
          schema:
            $ref: '#/definitions/ResultSetError'
        500:
          description: Server error
          schema:
            $ref: '#/definitions/ResultSetError'
    delete:
      description: "Revokes a specific release moving it back to initial Work status. A constraint is that there is no clashing entries in Work."
      summary: ReleaseRevoke
      tags:
      - release
      operationId: ReleaseRevoke
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        type: string
        format: uuid
        description: ''
      - name: release
        in: path
        required: true
        enum:
        - 0
        - last
        type: string
        description: release sequential number, 0 for work, else production
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/ResultSetOk'
          headers: {}
        401:
          description: Not authenticated
          schema:
            $ref: '#/definitions/ResultSetError'
        500:
          description: Server error
          schema:
            $ref: '#/definitions/ResultSetError'            
...
definitions:
...
  ResultSetError:
    title: ResultSetError
    type: object
    properties:
      status:
        $ref: '#/definitions/Status12'
      message:
        type: string
  ResultSetOk:
    title: ResultSetOk
    type: object
    properties:
      status:
        $ref: '#/definitions/Status'
...
  Status:
    title: Status
    example: Ok
    type: string
    enum:
    - Ok
  Status12:
    title: Status12
    example: Error
    type: string
    enum:
    - Error
...